ci(docx-core): eliminate import cycles + add madge --circular gate
Make the module graph acyclic and keep it that way:

- Break the document_view barrel back-edge (cycles 1-4): extract shared
  types into primitives/document_view-types.ts (a leaf module); the
  document_view-* submodules import types from there instead of from the
  document_view.js barrel. document_view.ts re-exports them, so its
  public surface is unchanged.
- Break the pipeline<->index barrel cycle (cycle 5): pipeline.ts and the
  compare/reconstruction types move to a leaf compare-types.ts; pipeline
  imports from defining modules instead of the ../../index.js barrel.
  index.ts re-exports the compare types, so the package surface is unchanged.
- Add `check:cycles` (madge --circular) script + a "Check for import
  cycles" step in the workspace-lint CI job + append to preflight:ci.
  madge@8.0.0 added as a root devDependency.

Behavior-preserving: generated .d.ts named-export sets of index.d.ts
(34) and document_view.d.ts (31) are byte-identical to origin/main;
type-only imports stay type-only. `npx madge --circular` now exits 0;
build + lint:workspaces + 1302 docx-core tests pass.

Closes #321.
