+++ package.json
@@ -19,6 +19,7 @@
     "build": "npm run build --workspaces --if-present",
     "lint:workspaces": "npm run lint --workspaces --if-present && npm run lint:allure-imports",
     "lint:allure-imports": "eslint \"packages/**/*.test.ts\" --ignore-pattern \"packages/google-docs-core/**\"",
+    "check:cycles": "madge --circular --extensions ts packages/*/src",
     "test": "npm run test --workspaces --if-present",
     "test:run": "npm run test:run --workspaces --if-present",
     "test:baseline": "npm run test:baseline -w @usejunior/docx-core",
@@ -42,7 +43,7 @@
     "check:allure-filenames": "node scripts/check_allure_test_filename_policy.mjs --baseline coverage/allure-test-filename-baseline.json",
     "check:mcpb-manifest": "npm run check:manifest-contract -w @usejunior/safedocx-mcpb",
     "check:gemini-extension-manifest": "node scripts/check_gemini_extension_manifest.mjs",
-    "preflight:ci": "npm run lint:workspaces && npm run check:spec-coverage && npm run check:tool-docs && npm run check:mcpb-manifest && npm run check:gemini-extension-manifest && npm run check:allure-labels && npm run check:allure-filenames && npm run check:allure-quality && npm run check:conformance-citations && npm run check:conformance-doc && npm run check:site",
+    "preflight:ci": "npm run lint:workspaces && npm run check:spec-coverage && npm run check:tool-docs && npm run check:mcpb-manifest && npm run check:gemini-extension-manifest && npm run check:allure-labels && npm run check:allure-filenames && npm run check:allure-quality && npm run check:conformance-citations && npm run check:conformance-doc && npm run check:site && npm run check:cycles",
     "test:coverage:packages": "npm run test:coverage -w @usejunior/docx-core && npm run test:coverage -w @usejunior/docx-mcp",
     "coverage:matrix": "node scripts/print_coverage_matrix.mjs",
     "coverage:packages:report": "node scripts/report_package_coverage.mjs --baseline coverage/package-coverage-baseline.json --output coverage/package-coverage-summary.json",
@@ -76,6 +77,7 @@
     "allure": "3.8.2",
     "eslint": "^10.0.1",
     "fast-xml-parser": "^5.7.0",
+    "madge": "8.0.0",
     "prismjs": "^1.30.0"
   },
   "overrides": {
